pyproject.toml
[project]
name = "agentmail-inbox-ops"
version = "0.1.0"
description = "Inbox operations scripts for AgentMail"
requires-python = ">=3.11"
dependencies = [
  "agentmail>=0.2.17",
  "python-dotenv>=1.2.1",
  "httpx>=0.28.1",
  "pypdf>=5.2.0",
  "python-docx>=1.1.2",
]
